wip: move i18n to state management
