+++ src/plugins/toolbar/components/BoldItalicUnderlineToggles.tsx
@@ -1,4 +1,4 @@
-import { useI18n } from '@/i18n/I18nProvider'
+import { useI18n } from '@/plugins/core/i18n'
 import { useCellValues, usePublisher } from '@mdxeditor/gurx'
 import React from 'react'
 import { IS_BOLD, IS_ITALIC, IS_UNDERLINE } from '../../../FormatConstants'
