+++ src/plugins/toolbar/components/DiffSourceToggleWrapper.tsx
@@ -3,7 +3,7 @@ import { useCellValues, usePublisher } from '@mdxeditor/gurx'
 import React from 'react'
 import styles from '../../../styles/ui.module.css'
 import { SingleChoiceToggleGroup } from '.././primitives/toolbar'
-import { useI18n } from '@/i18n/I18nProvider'
+import { useI18n } from '@/plugins/core/i18n'
 
 /**
  * A wrapper element for the toolbar contents that lets the user toggle between rich text, diff and source mode.
