+++ src/plugins/toolbar/components/InsertCodeBlock.tsx
@@ -1,4 +1,4 @@
-import { useI18n } from '@/i18n/I18nProvider'
+import { useI18n } from '@/plugins/core/i18n'
 import { useCellValue, usePublisher } from '@mdxeditor/gurx'
 import React from 'react'
 import { insertCodeBlock$ } from '../../codeblock/'
