+++ src/plugins/toolbar/components/InsertFrontmatter.tsx
@@ -1,4 +1,4 @@
-import { useI18n } from '@/i18n/I18nProvider'
+import { useI18n } from '@/plugins/core/i18n'
 import { useCellValues, usePublisher } from '@mdxeditor/gurx'
 import classNames from 'classnames'
 import React from 'react'
