+++ src/plugins/toolbar/components/InsertSandpack.tsx
@@ -3,7 +3,7 @@ import { ButtonOrDropdownButton } from '.././primitives/toolbar'
 import { insertSandpack$, sandpackConfig$ } from '../../sandpack'
 import { useCellValues, usePublisher } from '@mdxeditor/gurx'
 import { iconComponentFor$ } from '../../core'
-import { useI18n } from '@/i18n/I18nProvider'
+import { useI18n } from '@/plugins/core/i18n'
 
 /**
  * A dropdown button that allows the user to insert a live code block into the editor. The dropdown offers a list of presets that are defined in the sandpack plugin config.
